update Wrong code

+++ src/main/java/com/xiansi/dto/CommentDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 //把前端的数据反序列化
 @Data
 public class CommentDTO {
-	private Long parent_id;
+	private Integer parent_id;
 	private String content;
 	private Integer type;
 }
